+++ src/test/java/org/medicmobile/webapp/mobile/UtilsTest.java
@@ -1,17 +1,28 @@
 package org.medicmobile.webapp.mobile;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
 import android.net.Uri;
 
 import org.junit.*;
+import org.junit.rules.TemporaryFolder;
 import org.junit.runner.*;
 import org.robolectric.*;
 import org.robolectric.annotation.*;
 
-import static org.junit.Assert.*;
+import java.io.File;
+import java.io.IOException;
+import java.util.Optional;
 
 @RunWith(RobolectricTestRunner.class)
 @Config(sdk=28)
 public class UtilsTest {
+
+	@Rule
+	public TemporaryFolder temporaryFolder = new TemporaryFolder();
+
 	@Test
 	public void isUrlRelated_goodNormalUrls() {
 		final String[] goodUrls = {
@@ -50,6 +61,47 @@ public class UtilsTest {
 			assertFalse("Expected URL to be rejected, but it wasn't: " + badUrl,
 					Utils.isUrlRelated("https://example.com", Uri.parse(badUrl)));
 		}
+	}
+
+	@Test
+	public void getUriFromFilePath_withInvalidPath_returnsNull() {
+		//> WHEN
+		Optional<Uri> uriWithNullPath = Utils.getUriFromFilePath(null);
+		Optional<Uri> uriWithEmptyPath = Utils.getUriFromFilePath("");
+		Optional<Uri> uriMissingFile = Utils.getUriFromFilePath("/storage/file.txt");
+
+		//> THEN
+		assertFalse(uriWithNullPath.isPresent());
+		assertFalse(uriWithEmptyPath.isPresent());
+		assertFalse(uriMissingFile.isPresent());
+	}
+
+	@Test
+	public void getUriFromFilePath_withContentSchema_returnsUri() {
+		//> WHEN
+		Optional<Uri> uriOptional = Utils.getUriFromFilePath("content://folder/file.png");
+
+		//> THEN
+		assertTrue(uriOptional.isPresent());
+		Uri uri = uriOptional.get();
+		assertEquals("content", uri.getScheme());
+		assertEquals("/file.png", uri.getPath());
+	}
+
+	@Test
+	public void getUriFromFilePath_withFileSchema_returnsUri() throws IOException {
+		//> GIVEN
+		File file = temporaryFolder.newFile("some_file.txt");
+		String filePath = file.getPath();
+
+		//> WHEN
+		Optional<Uri> uriOptional = Utils.getUriFromFilePath(filePath);
+
+		//> THEN
+		assertTrue(uriOptional.isPresent());
+		Uri uri = uriOptional.get();
+		assertEquals("file", uri.getScheme());
+		assertEquals(filePath, uri.getPath());
 	}
 }