+++ src/webview/java/org/medicmobile/webapp/mobile/EmbeddedBrowserActivity.java
@@ -60,6 +60,8 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	static final int GRAB_PHOTO_ACTIVITY_REQUEST_CODE = (0 << 3) | NON_SIMPRINTS_FLAGS;
 	static final int GRAB_MRDT_PHOTO_ACTIVITY_REQUEST_CODE = (1 << 3) | NON_SIMPRINTS_FLAGS;
 	static final int DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE = (2 << 3) | NON_SIMPRINTS_FLAGS;
+	static final int ACCESS_STORAGE_PERMISSION_REQUEST_CODE = (3 << 3) | NON_SIMPRINTS_FLAGS;
+	static final int CHT_EXTERNAL_APP_ACTIVITY_REQUEST_CODE = (4 << 3) | NON_SIMPRINTS_FLAGS;
 
 	// Arbitrarily selected value
 	private static final int ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE = 7038678;
@@ -85,6 +87,7 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	private MrdtSupport mrdt;
 	private PhotoGrabber photoGrabber;
 	private SmsSender smsSender;
+	private ChtExternalAppHandler chtExternalAppHandler;
 
 	private boolean isMigrationRunning = false;
 
@@ -97,6 +100,7 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		this.simprints = new SimprintsSupport(this);
 		this.photoGrabber = new PhotoGrabber(this);
 		this.mrdt = new MrdtSupport(this);
+		this.chtExternalAppHandler = new ChtExternalAppHandler(this);
 		try {
 			this.smsSender = new SmsSender(this);
 		} catch(Exception ex) {
@@ -246,6 +250,9 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 							}
 						}
 						return;
+					case CHT_EXTERNAL_APP_ACTIVITY_REQUEST_CODE:
+						processChtExternalAppResult(resultCode, i);
+						return;
 					default:
 						trace(this, "onActivityResult() :: no handling for requestCode=%s",
 								requestCodeToString(requestCode));
@@ -262,16 +269,6 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		}
 	}
 
-	private String requestCodeToString(int requestCode) {
-		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
-			return "ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE";
-		}
-		if (requestCode == DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE) {
-			return "DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE";
-		}
-		return String.valueOf(requestCode);
-	}
-
 //> ACCESSORS
 	SimprintsSupport getSimprintsSupport() {
 		return this.simprints;
@@ -285,25 +282,24 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		return this.smsSender;
 	}
 
+	ChtExternalAppHandler getChtExternalAppLauncherActivity() {
+		return this.chtExternalAppHandler;
+	}
+
 //> PUBLIC API
 	public void evaluateJavascript(final String js) {
 		evaluateJavascript(js, true);
 	}
 
 	public void evaluateJavascript(final String js, final boolean useLoadUrl) {
-		container.post(() -> {
-			// `WebView.loadUrl()` seems to be significantly faster than
-			// `WebView.evaluateJavascript()` on Tecno Y4.  We may find
-			// confusing behaviour on Android 4.4+ when using `loadUrl()`
-			// to run JS, in which case we should switch to the second
-			// block.
-			// On switching to XWalkView, we assume the same applies.
-			if(useLoadUrl) { // NOPMD
-				container.loadUrl("javascript:" + js, null);
-			} else {
-				container.evaluateJavascript(js, IGNORE_RESULT);
-			}
-		});
+		int maxUrlSize = 2097100; // Maximum character limit supported for loading as url.
+
+		if (useLoadUrl && js.length() <= maxUrlSize) {
+			// `WebView.loadUrl()` seems to be significantly faster than `WebView.evaluateJavascript()` on Tecno Y4.
+			container.post(() -> container.loadUrl("javascript:" + js, null));
+		} else {
+			container.post(() -> container.evaluateJavascript(js, IGNORE_RESULT));
+		}
 	}
 
 	public void errorToJsConsole(String message, Object... extras) {
@@ -313,6 +309,24 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	}
 
 //> PRIVATE HELPERS
+	private String requestCodeToString(int requestCode) {
+		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+			return "ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE";
+		}
+
+		if (requestCode == DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE) {
+			return "DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE";
+		}
+
+		return String.valueOf(requestCode);
+	}
+
+	private void processChtExternalAppResult(int resultCode, Intent intentData) {
+		String script = this.chtExternalAppHandler.processResult(resultCode, intentData);
+		trace(this, "ChtExternalAppHandler :: Executing JavaScript: %s", script);
+		evaluateJavascript(script);
+	}
+
 	private void configureUseragent() {
 		String current = WebSettings.getDefaultUserAgent(this);
 		container.getSettings().setUserAgentString(createUseragentFrom(current));
@@ -402,10 +416,21 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	@Override
 	public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
 		super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-		if (requestCode != ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+		boolean granted = grantResults.length > 0 && grantResults[0] == PERMISSION_GRANTED;
+
+		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+			locationRequestResolved();
 			return;
 		}
-		locationRequestResolved();
+
+		if (requestCode == ACCESS_STORAGE_PERMISSION_REQUEST_CODE) {
+			if (granted) {
+				this.chtExternalAppHandler.resumeActivity();
+				return;
+			}
+			trace(this, "ChtExternalAppHandler :: User rejected permission.");
+			return;
+		}
 	}
 
 	public void locationRequestResolved() {