+++ src/xwalk/java/org/medicmobile/webapp/mobile/EmbeddedBrowserActivity.java
@@ -55,6 +55,8 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	static final int GRAB_PHOTO_ACTIVITY_REQUEST_CODE = (0 << 3) | NON_SIMPRINTS_FLAGS;
 	static final int GRAB_MRDT_PHOTO_ACTIVITY_REQUEST_CODE = (1 << 3) | NON_SIMPRINTS_FLAGS;
 	static final int DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE = (2 << 3) | NON_SIMPRINTS_FLAGS;
+	static final int ACCESS_STORAGE_PERMISSION_REQUEST_CODE = (3 << 3) | NON_SIMPRINTS_FLAGS;
+	static final int CHT_EXTERNAL_APP_ACTIVITY_REQUEST_CODE = (4 << 3) | NON_SIMPRINTS_FLAGS;
 
 	// Arbitrarily selected value
 	private static final int ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE = 7038678; // Arbitrarily selected value
@@ -80,6 +82,7 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	private MrdtSupport mrdt;
 	private PhotoGrabber photoGrabber;
 	private SmsSender smsSender;
+	private ChtExternalAppHandler chtExternalAppHandler;
 
 //> ACTIVITY LIFECYCLE METHODS
 	@Override public void onCreate(Bundle savedInstanceState) {
@@ -90,6 +93,7 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		this.simprints = new SimprintsSupport(this);
 		this.photoGrabber = new PhotoGrabber(this);
 		this.mrdt = new MrdtSupport(this);
+		this.chtExternalAppHandler = new ChtExternalAppHandler(this);
 		try {
 			this.smsSender = new SmsSender(this);
 		} catch(Exception ex) {
@@ -227,6 +231,9 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 							}
 						}
 						return;
+					case CHT_EXTERNAL_APP_ACTIVITY_REQUEST_CODE:
+						processChtExternalAppResult(resultCode, i);
+						return;
 					default:
 						trace(this, "onActivityResult() :: no handling for requestCode=%s",
 								requestCodeToString(requestCode));
@@ -243,16 +250,6 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		}
 	}
 
-	private String requestCodeToString(int requestCode) {
-		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
-			return "ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE";
-		}
-		if (requestCode == DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE) {
-			return "DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE";
-		}
-		return String.valueOf(requestCode);
-	}
-
 //> ACCESSORS
 	SimprintsSupport getSimprintsSupport() {
 		return this.simprints;
@@ -266,15 +263,20 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 		return this.smsSender;
 	}
 
+	ChtExternalAppHandler getChtExternalAppHandler() {
+		return this.chtExternalAppHandler;
+	}
+
 //> PUBLIC API
 	public void evaluateJavascript(final String js) {
-		// `WebView.loadUrl()` seems to be significantly faster than
-		// `WebView.evaluateJavascript()` on Tecno Y4.  We may find
-		// confusing behaviour on Android 4.4+ when using `loadUrl()`
-		// to run JS, in which case we should switch to the second
-		// block.
-		// On switching to XWalkView, we assume the same applies.
-		container.post(() -> container.load("javascript:" + js, null));
+		int maxUrlSize = 2097100; // Maximum character limit supported for loading as url.
+
+		if (js.length() <= maxUrlSize) {
+			// `loadUrl()` seems to be significantly faster than `evaluateJavascript()` on Tecno Y4.
+			container.post(() -> container.load("javascript:" + js, null));
+		} else {
+			container.post(() -> container.evaluateJavascript(js, IGNORE_RESULT));
+		}
 	}
 
 	public void errorToJsConsole(String message, Object... extras) {
@@ -286,6 +288,24 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 	}
 
 //> PRIVATE HELPERS
+	private String requestCodeToString(int requestCode) {
+		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+			return "ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE";
+		}
+
+		if (requestCode == DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE) {
+			return "DISCLOSURE_LOCATION_ACTIVITY_REQUEST_CODE";
+		}
+
+		return String.valueOf(requestCode);
+	}
+
+	private void processChtExternalAppResult(int resultCode, Intent intentData) {
+		String script = this.chtExternalAppHandler.processResult(resultCode, intentData);
+		trace(this, "ChtExternalAppHandler :: Executing JavaScript: %s", script);
+		evaluateJavascript(script);
+	}
+
 	private void jsConsole(String type, String message, Object... extras) {
 		String formatted = String.format(message, extras);
 		String escaped = formatted.replace("'", "\\'");
@@ -390,10 +410,21 @@ public class EmbeddedBrowserActivity extends LockableActivity {
 
 	@Override
 	public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-		if (requestCode != ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+		boolean granted = grantResults.length > 0 && grantResults[0] == PERMISSION_GRANTED;
+
+		if (requestCode == ACCESS_FINE_LOCATION_PERMISSION_REQUEST_CODE) {
+			locationRequestResolved();
 			return;
 		}
-		locationRequestResolved();
+
+		if (requestCode == ACCESS_STORAGE_PERMISSION_REQUEST_CODE) {
+			if (granted) {
+				this.chtExternalAppHandler.resumeActivity();
+				return;
+			}
+			trace(this, "ChtExternalAppHandler :: User rejected permission.");
+			return;
+		}
 	}
 
 	public void locationRequestResolved() {